Separating nodes in their own catalogs
