+++ package.json
@@ -26,6 +26,7 @@
     "react-dom": "^18.3.1",
     "react-i18next": "^15.0.1",
     "react-use-draggable-scroll": "^0.4.7",
+    "zod": "^4.4.3",
     "zustand": "^4.5.5"
   },
   "devDependencies": {
