+++ src/features/logic/nodes/shared/NodePanel.tsx
