+++ src/features/logic/nodes/shared/NodeRichMessage.tsx
