+++ src/features/logic/nodes/type/open/NodeOpen.tsx
@@ -1,0 +1,36 @@
+import { NodeProps, Position } from "@xyflow/react";
+import NodePanel from "../../shared/NodePanel";
+import { NodeHandle } from "../../shared/NodeHandle";
+import NodeHeader from "../../shared/NodeHeader";
+import NodeOptions from "../../shared/NodeOptions";
+import { useTranslation } from "react-i18next";
+import NodeRichMessage from "../../shared/NodeRichMessage";
+import { TypeNode } from "./setup";
+import { FlowSelect } from "@/features/logic/components/controls/FlowSelect/FlowSelect";
+
+type Props = NodeProps<TypeNode>;
+
+export function NodeOpen({ id, type, selected, data }: Props) {
+  const { t } = useTranslation();
+  //   const onChange = useCallback((evt) => {
+  //     console.log(evt.target.value);
+  //   }, []);
+
+  return (
+    <NodePanel className="" isSelected={selected}>
+      <NodeHandle type="target" position={Position.Top} />
+      <NodeHeader
+        type={type}
+        label={data.label}
+        // description={data.description}
+      />
+      <NodeOptions isSelected={selected}></NodeOptions>
+
+      <NodeRichMessage>
+        Open <FlowSelect value={["Google"]} setValue={() => {}} />
+        with
+        <FlowSelect value={["{phrase}"]} setValue={() => {}} />.
+      </NodeRichMessage>
+    </NodePanel>
+  );
+}
