+++ src/features/logic/nodes/type/recipe/setup.ts
@@ -1,0 +1,25 @@
+import { type Node } from "@xyflow/react";
+import z from "zod";
+import { nodeSharedDataSchema, shoukaiSearchEngineSchema } from "../schema";
+
+export const nodeSchema = nodeSharedDataSchema.extend({
+  search: z.object({
+    phrase: z.string(),
+    domain: z.string(),
+    engine: shoukaiSearchEngineSchema,
+  }),
+});
+
+export type TypeNodeData = z.infer<typeof nodeSchema>;
+
+export type TypeNode = Node<TypeNodeData, "recipe">;
+
+export const defaultData: TypeNode["data"] = {
+  label: "Recipe",
+  description: "Recipe",
+  search: {
+    phrase: "",
+    domain: "",
+    engine: "default",
+  },
+};
