+++ src/features/logic/nodes/type/schema.ts
@@ -1,0 +1,17 @@
+import z from "zod";
+import { ShoukaiSearchEngine } from "@/constants";
+
+export const nodeSharedDataSchema = z.object({
+  label: z.string(),
+  description: z.string(),
+});
+
+export type NodeSharedData = z.infer<typeof nodeSharedDataSchema>;
+
+export const shoukaiSearchEngineSchema = z.enum([
+  "default",
+  "bing",
+  "duckduckgo",
+  "google",
+  "yandex",
+] satisfies ShoukaiSearchEngine[]);
