+++ src/features/logic/nodes/type/start/setup.ts
@@ -1,0 +1,17 @@
+import { type Node } from "@xyflow/react";
+import z from "zod";
+import { nodeSharedDataSchema } from "../schema";
+
+export const nodeSchema = nodeSharedDataSchema.extend({
+  shortcuts: z.array(z.string()),
+});
+
+export type TypeNodeData = z.infer<typeof nodeSchema>;
+
+export type TypeNode = Node<TypeNodeData, "start">;
+
+export const defaultData: TypeNode["data"] = {
+  label: "Start",
+  description: "All new queries start here.",
+  shortcuts: ["d", "g", "img", "gm"],
+};
