+++ src/features/logic/stores/cache/getNodeFromCacheIfPossible.tsx
@@ -1,4 +1,4 @@
-import { ShoukaiNode } from "../../types";
+import { ShoukaiNode } from "../../components/flow/nodes/type/types";
 import { DiagramStore } from "../useDiagramStore";
 
 type Params = { nodeId: string; store: DiagramStore };
