+++ src/features/logic/stores/useDragAndDropStore.ts
@@ -1,23 +1,10 @@
 import { create } from "zustand";
 import { type XYPosition } from "@xyflow/react";
-import { ShoukaiNode, ShoukaiNodeType } from "../types";
+
 import { generateId } from "@/utils/ids";
 import { addNode } from "./useDiagramStore";
-// import type { SupportedNodeTypes } from "@ts";
-// import { generateId } from "@/utils/generateId";
-// import { addNode } from "@/versions/diagram-4-zustand-actions/stores/useGlobalStore";
-
-import { defaultData as defaultDataNodeStart } from "../components/flow/nodes/NodeStart";
-import { defaultData as defaultDataNodeRecipe } from "../components/flow/nodes/NodeRecipe";
-import { defaultData as defaultDataNodeOpen } from "../components/flow/nodes/NodeOpen";
-import { defaultData as defaultDataNodeIf } from "../components/flow/nodes/NodeIf";
-
-const defaultDataByNode: Record<ShoukaiNodeType, ShoukaiNode["data"]> = {
-  start: defaultDataNodeStart,
-  recipe: defaultDataNodeRecipe,
-  open: defaultDataNodeOpen,
-  if: defaultDataNodeIf,
-};
+import { defaultDataByNodeType } from "../nodes/type/nodeTypes";
+import { ShoukaiNode, ShoukaiNodeType } from "../nodes/type/types";
 
 export type DragAndDropStoreState = {
   draggedType: ShoukaiNodeType | null;
@@ -46,7 +33,7 @@ export const addNodeOnDropIfDragged = (position: XYPosition) => {
     id: generateId(draggedType),
     type: draggedType,
     position,
-    data: defaultDataByNode[draggedType],
+    data: defaultDataByNodeType[draggedType],
   } as ShoukaiNode;
 
   addNode(newNode);
